requirements.txt
flask
langgraph
langchain-anthropic
langchain-core
pandas
openpyxl
python-dotenv
pillow 
